Added

+++ Assets/Scripts/enemyScript/behaviour/cooldown.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class cooldown : StateMachineBehaviour
 {
     float timer;
-    public int timeIdle;
+    public float timeIdle;
     
     // OnStateEnter is called when a transition starts and the state machine starts to evaluate this state
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
